+++ arch/risc-v/src/esp32c3/Make.defs
@@ -174,7 +174,7 @@ endif
 
 ifeq ($(CONFIG_ESP32C3_WIRELESS),y)
 WIRELESS_DRV_UNPACK  = esp-wireless-drivers-3rdparty
-WIRELESS_DRV_ID      = 024420d
+WIRELESS_DRV_ID      = 055f1ef
 WIRELESS_DRV_ZIP     = $(WIRELESS_DRV_ID).zip
 WIRELESS_DRV_URL     = https://github.com/espressif/esp-wireless-drivers-3rdparty/archive
 
@@ -204,6 +204,10 @@ EXTRA_LIBS += -lphy
 # unused functions.
 
 LDFLAGS += --gc-sections
+
+# Wireless interfaces.
+
+CHIP_CSRCS += esp32c3_wireless.c
 endif
 
 ifeq ($(CONFIG_ESP32C3_WIFI),y)
