+++ boards/risc-v/esp32c3/esp32c3-devkit/scripts/esp32c3_rom.ld
@@ -1591,7 +1591,6 @@ ppTask = 0x40001720;
 ppTxPkt = 0x40001724;
 ppTxProtoProc = 0x40001728;
 ppTxqUpdateBitmap = 0x4000172c;
-pp_coex_tx_request = 0x40001730;
 pp_hdrsize = 0x40001734;
 pp_post = 0x40001738;
 pp_process_hmac_waiting_txq = 0x4000173c;
@@ -1762,12 +1761,10 @@ esp_coex_rom_version_get = 0x400018ac;
 coex_bt_release = 0x400018b0;
 coex_bt_request = 0x400018b4;
 coex_core_ble_conn_dyn_prio_get = 0x400018b8;
-coex_core_event_duration_get = 0x400018bc;
 coex_core_pti_get = 0x400018c0;
 coex_core_release = 0x400018c4;
 coex_core_request = 0x400018c8;
 coex_core_status_get = 0x400018cc;
-coex_core_timer_idx_get = 0x400018d0;
 coex_event_duration_get = 0x400018d4;
 coex_hw_timer_disable = 0x400018d8;
 coex_hw_timer_enable = 0x400018dc;
